Do not force class IoU calculation to be uint16 but uint8 in instance segmentation

+++ biapy/engine/instance_seg.py
@@ -685,8 +685,8 @@ class Instance_Segmentation_Workflow(Base_Workflow):
 
                 # Measure class IoU
                 class_iou = self.jaccard_index_matching(
-                    torch.as_tensor(class_channel.squeeze().astype(np.uint16)).to(self.test_device, non_blocking=True),
-                    torch.as_tensor(_Y_classes.squeeze().astype(np.uint16)).to(self.test_device, non_blocking=True),
+                    torch.as_tensor(class_channel.squeeze().astype(np.uint8)).to(self.test_device, non_blocking=True),
+                    torch.as_tensor(_Y_classes.squeeze().astype(np.uint8)).to(self.test_device, non_blocking=True),
                 )
                 class_iou = class_iou.item() if not torch.isnan(class_iou) else 0
                 print(f"Class IoU: {class_iou}")
